requirements.txt
python-dotenv
cohere
streamlit
pandas